+++ src/main/java/es/mira/progesin/persistence/repositories/IDocumentacionPreviaRepository.java
@@ -8,6 +8,6 @@ import es.mira.progesin.persistence.entities.DocumentacionPrevia;
 
 public interface IDocumentacionPreviaRepository extends CrudRepository<DocumentacionPrevia, String> {
 
-	List<DocumentacionPrevia> findByIdSolicitud(Integer idSolicitud);
+	List<DocumentacionPrevia> findByIdSolicitud(Long idSolicitud);
 
 }
